Removed encodeURI from facebooker.js
Having it there made my code fail, and I couldn't see what it was needed for?
Is there concern that FB's serialize() isn't sanitized enough or...?

How it failed for me was using a form field such as 'name="posts[subject]"' as the Form.serialize saw posts and passed the object which contained subject (and another posts variable) and didn't know how to handle that.